+++ infinigen/assets/scatters/ground_twigs.py
@@ -46,3 +46,37 @@ def apply(obj, selection=None, n_leaf=0, n_twig=10, **kwargs):
     )
 
     return scatter_obj, twigs
+
+
+class GroundTwigs:
+    def __init__(self):
+        pass
+
+    def apply(self, obj, selection=None, n_leaf=0, n_twig=10, **kwargs):
+        (_, twig_params, leaf_params), _ = random_species(season="winter")
+        twigs = make_twig_collection(
+            np.random.randint(1e5),
+            twig_params,
+            leaf_params,
+            n_leaf=n_leaf,
+            n_twig=n_twig,
+            leaf_types=None,
+            trunk_surface=weighted_sample(material_assignments.bark),
+        )
+
+        toggle_show(twigs)
+        for o in twigs.objects:
+            approx_settle_transform(o, samples=40)
+        toggle_hide(twigs)
+
+        scatter_obj = scatter_instances(
+            base_obj=obj,
+            collection=twigs,
+            scale=U(0.15, 0.3),
+            scale_rand=U(0, 0.3),
+            scale_rand_axi=U(0, 0.2),
+            density=10,
+            ground_offset=0.05,
+            selection=selection,
+            taper_density=True,
+        )